Updated TicTacToe and Checkers Plus new GameEngine
This update includes a fix for tic tac toe bug that changed the turn when an invalid move was made. It includes a new folder with abstract classes called game engine that includes abstract classes that can be extended by all games. My current version of tic tac toe and checkers use this as a base and so they share the same structure.