Fixed the invalid turn bug from freezing the game.
There is still a bug where if you click a single piece multiple times in a row. I haven't figured out the exact causes behind the bug, but if you click on the same piece three times and then attempt to move, the game will function normally. If you click four or more times, it seems to break. More testing is required (alternatively, we can hope the grader doesn't spam click the game).